requirements.txt
torch==2.6.0
torchvision==0.21.0
pillow==10.4.0
numpy==1.26.4
scikit-learn==1.3.2
